Allowing a generator per class
