+++ plugin/src/main/kotlin/com/github/gmazzo/buildconfig/generators/BuildConfigGenerator.kt
@@ -1,5 +1,12 @@
 package com.github.gmazzo.buildconfig.generators
 
 import org.gradle.api.Action
+import org.gradle.api.tasks.Input
 
-interface BuildConfigGenerator : Action<BuildConfigGeneratorSpec>
+@JvmDefaultWithoutCompatibility
+interface BuildConfigGenerator : Action<BuildConfigGeneratorSpec> {
+
+    @get:Input
+    val type: Class<out BuildConfigGenerator> get() = this::class.java
+
+}
